+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 2.4.0 - 14.02.2020
+* added option to ignore query clients allover
+* when query events are getting triggered it should now check if all required entities to fire the final event from the TeamSpeak Library otherwise it will throw an `EventError`
+
 ### 2.3.3 - 27.01.2020
 * removed channel_cpid
 * added cpid to channeledit props